Replaces `dbSelect` function types with interfaces.
Prior to this commit the `SelectFoo()` functions accepted a function
type as the first argument, and invoked it on behalf of the caller with
the constructed query. This might allow passing the wrong function to
`SelectFoo` (e.g. `Exec` instead of `SelectOne`).

Per feedback from @jsha this commit replaces these function types with
interfaces defining the required functions. This should be safer
overall.