Moves `mailer.Connect()` closer to `mailer.SendMail()`.
Per #2250 when we connect to the remote SMTP server at start-up time by
calling `mailer.Connect()` but do not actually call `mailer.SendMail()`
until after we have done some potentially expensive/time-consuming work
we are liable to have our connection closed due to timeout.

This commit moves the `Connect()` call in `expiration-mailer` and
`notify-mailer` to be closer to where the actual messages are sent via
`SendMail()`.