Aborts `NewAuthorization` when SA RPC fails.
The RA performs an RPC to the SA's `GetValidAuthorizations` function
when attempting to find existing valid authorizations to reuse. Prior to
this commit, ff the RPC fails (e.g. due to a timeout) the calling code
logs the failure as a warning but fails to return the error and cease
processing. This results in a nil panic when we later try to index
`auths`

This commit inserts the missing `return` to ensure we don't process
further, thereby resolving #2274.