Moves nil `serverConfig` check, adds TODO.
This commit moves the check for a nil `serverConfig` from
`ServerHandshake()` into `NewServerTransport()`. The encapsulation of
the non-exported field means this is the safer option.

This commit also adds a TODO related to the "fail open" behaviour where
there is no accepted SAN list configured. This should be removed once we
have configured each of the gRPC servers with a accepted SAN list.