Logs failed SA.UpdatePendingAuthorization.
This commit resolves #2303 by updating the comment, and returned error
type produced when the RA calls `SA.UpdatePendingAuthorization` and it
fails.

Previously this produced a `MalformedRequestError` that was
described as only happening when the client corrupted the challenge
data.

Now this is returned as more descriptive `ServerInternalError` and the
underlying error from the SA is logged as a warning for further
debugging.