Updates `go-statsd-client` to 2.0.2. (#2343)
This commit updates the cactus `go-statsd-client` to the 2.0.2 release
tag.

Note: the most recent release is 3.1.0 but the introduction of
substatter support in 3.0.0 changes the `Statter` interface we rely on
in the boulder metrics package. We should consider follow-up work to
refactor our metrics code to be compatible with the 3.x releases.

Relates to #2315.