Remove /go/bin from PATH in Docker images (#2412)
We don't write any built binaries there during the test process (Boulder
binaries go in boulder/bin), so including it in that path has no effect,
at least on Travis.

For local builds, because GOPATH gets mounted as a volume at /go/,
including /go/bin in PATH means that a docker run command locally can
wind up running a local copy of some key binary that is different than
the one that gets run in Travis. For instance, this happens when the
host's version of protoc-gen-go differs from the version in the image,
producing diffs when Travis runs the "generate" phase of test.sh.