Fixes the `config/publisher.json` clientNames list. (#2466)
In https://github.com/letsencrypt/boulder/pull/2453 we created
individual client certificates for each gRPC client. The "clientNames"
list in the `config-next/publisher.json` was updated for the new
component-specific SANs but we neglected to updated
`config/publisher.json`. This caused the `ocsp-updater` (which uses gRPC
in the base `config/` to talk to the `publisher`) to fail to connect.

This commit updates `config/publisher.json` to have the same clientNames
as `config-next/publisher.json` and resolves #2465