Mentions the non-default challenge ports in README.
By default the Boulder dev env's VA config `test/config/va.json` has
a `portConfig` that connects to port 5002 for HTTP-01 challenges and
port 5001 for TLS-SNI-01 challenges.

This can be confusing to someone that follows the FAKE_DNS guidelines
for using a client from the host machine but expects the VA to reach out
  on port 80 or 443 like production/staging's VA.

This commit updates the README to include a note on the non-standard
ports and how to change them.