Handle v4 db file correctly, configure logger.
Prior to this commit `gsb.go` was misconfigured to give a directory
filepath to the V4 api for its database usage. The library expects
a file to be given and the file must exist. This commit updates the
calling code to properly handle this.

An adapter is added to allow the v4 safebrowsing library to log to the
Boulder logger at INFO level.

Finally, the ServerURL and logger are provided to the v4 library
constructor.