Use X.509 AIA Issuer URL in rel="up" link header
In order to provide the correct issuer certificate for older
certificates after an issuer certificate rollover or when using multiple
issuer certificates (e.g. RSA and ECDSA), use the AIA CA Issuer URL
embedded in the certificate for the rel="up" link served by WFE.
This behaviour is gated behind the "UseAIAIssuerURL" feature,
which defaults to false.

To prevent MitM vulnerabilities in cases where the AIA URL is
HTTP-only, it is upgraded to HTTPS.

This also adds a test for the issuer URL returned by the /acme/cert
endpoint. wfe/test/178.{crt,key} were regenerated to add the AIA
extension required to pass the test.

/acme/cert was changed to return an absolute URL to the
issuer endpoint (making it consistent with /acme/new-cert).

Fixes #1663