Removes CertStatusOptimizationsMigrated Feature Flag.
The `NotAfter` and `IsExpired` fields on the `certificateStatus` table
have been migrated in staging & production. Similarly the
`CertStatusOptimizationsMigrated` feature flag has been turned on after
a successful backfill operation. We have confirmed the optimization is
working as expected and can now clean out the duplicated `v1` and `v2`
models, and the feature flag branching. The `notafter-backfill` command
is no longer useful and so this commit also cleans it out of the repo.

Note: Some unit tests were sidestepping the SA and inserting
`certificateStatus` rows explicitly. These tests had to be updated to
set the `NotAfter` field in order for the queries used by the
ocsp-updater and the expiration-mailer to perform the way the tests
originally expected.