Fix missing berrors.InvalidEmail -> probs.ProblemDetails mapping (#2630)
This fixes an issue caused by #2583. Prior to that PR, we would serve the "invalidEmail" problem type when a DNS lookup for an email base domain failed. After that PR, we would map "berrors.InvalidEmail" to the "InternalServerError" problem type, which caused 500 errors to be returned to the user.

This PR restores the behavior of returning "type": "...invalidEmail" to the user.