CA: Orphan certificate during issuance if it can't be parsed.
It is better to orphan a malformed certificate than store it in the
database where it could somehow be retrieved. Such a malformed
certificate would already have been incorporated into the audit log
at the point we orphan it, so there is a record of it. In particular
there is no point in sending a malformed certificate to CT logs.

For further motivation, it is expected that we'll soon need to
use the parsed certificate later in the function.