CA: Don't add cert DER to audit log for OCSP failures. (#2667)
First, commit c0ad8d90403cb8a965254ade56c902f4810fec94 (PR #2658)
had a minor bug: It didn't update the "pem=" in the audit log
message to "cert=" to be consistent with the rest of the code.

But, more importantly, we don't need to include the cert DER in the
audit log at this point because we've already logged the DER and its
serial number prior to this. Thus, at this point logging the serial
number is good enough.