Fixes RPC bug for CountCertificatesExact feature flag.
With the `CountCertificatesExact` feature flag enabled if the RA's
`checkCertificatesPerNameLimit` was called with `names` only containing
domains exactly matching a public suffix entry then the legacy
`ra.enforceNameCounts` function will be called with an empty `tldNames`
argument. This in turn will cause the RA->SA RPC to fail with an
"incomplete gRPC request message error".

This commit fixes this bug by only calling `ra.enforceNameCounts` when
`len(tldNames) > 0`.

Since the bug happened at the RPC marshalling error it isn't possible
to catch with a unit test in `ra_test.go` (e.g.
TestExactPublicSuffixCertLimit) since it uses a mock SA that doesn't use
gRPC. As a result this commit adds a brief integration test for a public
suffix domain issuance. Before this commit this integration test failed
with the incomplete gRPC error. Afterwards it works. Yay!