Fix TestGetFQDNSetsBySerials flakyness.
This commit fixes the SA `TestGetFQDNSetsBySerials` test to avoid flaky
failures from assuming that the results will be ordered. The fix is to
track the expected hashes with the count of times we expect to see them
in the results as a map. Each time we see a result we decrement the
count, deleting the map key when the count reaches 0. At the end of the
test if all expected values were seen the map will be empty.