Update README for RAM requirement, add `cd` command.
From... ahem... some frustrating debugging I determined that the Boulder
docker environment fails in strange & mysterious ways if you do not have
sufficient RAM. This commit adds this fact to the README to save future
souls my torment.

This commit also adds a `cd` to the intial `git clone` instructions to
ensure the user is in the correct directory to run `docker-compose up`
from.